refactor: move point to geometry directory
