+++ src/geometry/Point.test.ts
