+++ src/geometry/Point.ts
